+++ src/main/java/com/investing/start/services/CapRateService.java
@@ -9,8 +9,8 @@ public class CapRateService {
     public CapRateService(NOIService noiService) {
         this.noiService = noiService;
     }
-    public int getCapRate(House house) {
+    public double getCapRate(House house) {
         int noi = noiService.getNOI(house);
-        return (noi / house.homeValue()) * 100;
+        return ((double) noi / house.homeValue()) * 100;
     }
 }
